Module ec_gpu_gen::multiexp_cpu

Multiexponentiation on the CPU.

Structs

DensityTracker

FullDensity

Traits

QueryDensity

Source

A source of bases, like an iterator.

SourceBuilder

An object that builds a source of bases.

Functions

multiexp_cpu

Perform multi-exponentiation. The caller is responsible for ensuring the query size is the same as the number of exponents.